requirements.txt
streamlit>=1.50.0
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
peft>=0.4.0
mediapipe>=0.10.0
opencv-python>=4.8.0
pillow>=9.5.0
numpy>=1.24.0
requests>=2.31.0





